Перенастроил логику работы CoreData, провел рефакторинг кода, добавил иконку приложению

+++ SimpleProject/StockScreen/StockViewController.swift
@@ -17,51 +17,63 @@ class StockViewController: UIViewController {
         checkCoreData()
     }
     
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(true)
+        
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         stockModel.loadInfo()
         stockModel.loadPrice()
         stockModel.stockInfoDidChange = {
             DispatchQueue.main.async { [weak self] in
-                self?.tickerLabel.text = self?.stockModel.stockInfo.first?.ticker
-                self?.nameLabel.text = self?.stockModel.stockInfo.first?.name
-                self?.descriptionLabel.text = self?.stockModel.stockInfo.first?.description
+                self?.configureStockInfo()
             }
         }
+        
         stockModel.stockPriceDidChange = {
             DispatchQueue.main.async { [weak self] in
-                self?.closePrice.text = String(self?.stockModel
-                    .stockPrice.first?.close ?? 0.00)
-                self?.openPrice.text = String(self?.stockModel
-                    .stockPrice.first?.open ?? 0.00)
-                let difference = self?.stockModel.calcDifference() ?? 0.00
-                if difference > 0 {
-                    self?.differencePrice.textColor = .green
-                    self?.differencePrice.text = "+" + String(format: "%.2f", difference)
-                } else {
-                    self?.differencePrice.textColor = .red
-                    self?.differencePrice.text = String(format: "%.2f", difference)
-                }
+                self?.configurePriceInfo()
             }
         }
     }
     
+    private func configureStockInfo() {
+        tickerLabel.text = stockModel.stockInfo.first?.ticker
+        nameLabel.text = stockModel.stockInfo.first?.name
+        descriptionLabel.text = stockModel.stockInfo.first?.description
+    }
+    
+    private func configurePriceInfo() {
+        closePrice.text = String(self.stockModel
+            .stockPrice.first?.close ?? 0.00)
+        openPrice.text = String(self.stockModel
+            .stockPrice.first?.open ?? 0.00)
+        let difference = stockModel.calcDifference()
+        if difference > 0 {
+            differencePrice.textColor = .green
+            differencePrice.text = "+" + String(format: "%.2f", difference)
+        } else {
+            differencePrice.textColor = .red
+            differencePrice.text = String(format: "%.2f", difference)
+        }
+    }
+    
     private func checkCoreData() {
-        
         let request = Stock.fetchRequest()
         if let stocks = try? CoreDataService.context.fetch(request) {
             if stocks.isEmpty {
                 state = .add
             } else {
-                for stock in stocks {
-                    if stock.stockName == self.stockModel.stock {
-                        state = .delete
-                        addStockButton.setImage(UIImage(systemName: "heart.fill"), for: .normal)
-                    } else {
-                        state = .add
-                        addStockButton.setImage(UIImage(systemName: "heart"), for: .normal)
-                    }
-                }    
+                let stockNames = stocks.map { $0.stockName }
+                if stockNames.contains(self.stockModel.stock) {
+                    state = .delete
+                    addStockButton.setImage(UIImage(systemName: "heart.fill"), for: .normal)
+                } else {
+                    state = .add
+                    addStockButton.setImage(UIImage(systemName: "heart"), for: .normal)
+                }
             }
         }
     }
@@ -72,21 +84,24 @@ class StockViewController: UIViewController {
             let request = Stock.fetchRequest()
             if let stocks = try? CoreDataService.context.fetch(request) {
                 for stock in stocks {
-                    CoreDataService.context.delete(stock)
-                    CoreDataService.saveContext()
-                    addStockButton.setImage(UIImage(systemName: "heart"), for: .normal)
+                    if stock.stockName == self.stockModel.stock {
+                        CoreDataService.context.delete(stock)
+                        CoreDataService.saveContext()
+                        addStockButton.setImage(UIImage(systemName: "heart"), for: .normal)
+                    }
                 }
             }
         case .add :
             let context = CoreDataService.context
             context.perform {
                 let newStock = Stock(context: context)
+                newStock.openPrice = self.stockModel
+                    .stockPrice.first?.open ?? 0.00
+                newStock.differencePrice = self.stockModel.calcDifference()
                 newStock.stockName = self.stockModel.stock
                 CoreDataService.saveContext()
             }
             addStockButton.setImage(UIImage(systemName: "heart.fill"), for: .normal)
-            
-            
         case .none:
             break
         }
